requirements.txt
tensorflow
flask
opencv-python
numpy
pillow
